+++ src/main/java/com/mageddo/dnsproxyserver/di/module/ModuleStartup.java
@@ -2,7 +2,7 @@ package com.mageddo.dnsproxyserver.di.module;
 
 import com.mageddo.dnsproxyserver.di.StartupEvent;
 import com.mageddo.dnsproxyserver.dnsconfigurator.DnsConfigurators;
-import com.mageddo.dnsproxyserver.docker.EventListener;
+import com.mageddo.dnsproxyserver.server.dns.solver.docker.entrypoint.EventListener;
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ElementsIntoSet;
