+++ src/main/java/com/mageddo/dnsproxyserver/docker/dataprovider/ContainerFacadeDefault.java
@@ -1,8 +1,9 @@
-package com.mageddo.dnsproxyserver.docker;
+package com.mageddo.dnsproxyserver.docker.dataprovider;
 
 import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.api.command.InspectContainerResponse;
 import com.github.dockerjava.api.model.Container;
+import com.mageddo.dnsproxyserver.docker.application.Containers;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
