+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/application/DockerNetworkService.java
@@ -1,7 +1,9 @@
-package com.mageddo.dnsproxyserver.docker;
+package com.mageddo.dnsproxyserver.server.dns.solver.docker.application;
 
 import com.github.dockerjava.api.model.Container;
 import com.github.dockerjava.api.model.Network;
+import com.mageddo.dnsproxyserver.docker.dataprovider.ContainerFacade;
+import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacade;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.ContainerDAO;
 import com.mageddo.net.IP;
 import lombok.RequiredArgsConstructor;
@@ -12,9 +14,9 @@ import javax.inject.Singleton;
 import java.util.ArrayList;
 import java.util.List;
 
-
 import static com.mageddo.dnsproxyserver.server.dns.solver.docker.Network.Name;
 
+// fixme #444 acoplado a infraestrutura
 @Slf4j
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__({@Inject}))
