+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/dataprovider/DockerNetworkDAODefault.java
@@ -1,7 +1,7 @@
 package com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider;
 
 import com.github.dockerjava.api.model.Container;
-import com.mageddo.dnsproxyserver.docker.DockerNetworkFacade;
+import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacade;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.Network;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.mapper.NetworkMapper;
 import lombok.AllArgsConstructor;
