+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/dataprovider/mapper/DpsContainer.java
@@ -1,4 +1,4 @@
-package com.mageddo.dnsproxyserver.docker;
+package com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.mapper;
 
 import com.mageddo.dnsproxyserver.server.dns.Hostname;
 import com.mageddo.dnsproxyserver.utils.Splits;
