+++ src/test/java/com/mageddo/dnsproxyserver/docker/DockerConnectionCheckTest.java
@@ -4,7 +4,8 @@ import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.api.command.VersionCmd;
 import com.mageddo.commons.concurrent.Threads;
 import com.mageddo.dnsproxyserver.di.module.ModuleDockerClient;
-import com.mageddo.dnsproxyserver.docker.DockerConnectionCheck.Status;
+import com.mageddo.dnsproxyserver.docker.application.DockerConnectionCheck;
+import com.mageddo.dnsproxyserver.docker.application.DockerConnectionCheck.Status;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
