+++ src/test/java/com/mageddo/dnsproxyserver/docker/DockerNetworkFacadeDefaultTest.java
@@ -2,6 +2,7 @@ package com.mageddo.dnsproxyserver.docker;
 
 import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.api.exception.DockerException;
+import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacadeDefault;
 import testing.templates.docker.DockerClientTemplates;
 import com.mageddo.http.HttpStatus;
 import org.junit.jupiter.api.BeforeEach;
