+++ src/test/java/com/mageddo/dnsproxyserver/docker/DpsContainerManagerTest.java
@@ -1,12 +1,13 @@
 package com.mageddo.dnsproxyserver.docker;
 
+import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.DpsContainerManager;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static com.mageddo.dnsproxyserver.docker.DpsContainerManager.DPS_INSIDE_CONTAINER;
+import static com.mageddo.dnsproxyserver.server.dns.solver.docker.application.DpsContainerManager.DPS_INSIDE_CONTAINER;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.doReturn;
